requirements.txt
pillow==11.1.0
pyautogui==0.9.54
pyautoit==0.6.5
